test(favorite): move add favorite tests from UserControllerTest to FavoriteControllerTest
# Conflicts:
#	itachallenge-user/src/test/java/com/itachallenge/user/controller/userinteraction/favorite/FavoriteControllerTest.java
