+++ itachallenge-user/src/test/java/com/itachallenge/user/controller/userinteraction/favorite/FavoriteControllerTest.java
@@ -39,6 +39,86 @@ class FavoriteControllerTest {
     }
 
     @Test
+    void addToFavorites_WhenAdded_Returns201() {
+        String userId = UUID.randomUUID().toString();
+        String challengeId = UUID.randomUUID().toString();
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
+                .thenReturn(Mono.just(true));
+
+        webTestClient.post()
+                .uri("/itachallenge/api/v1/user/users/" + userId + "/favorites/" + challengeId)
+                .exchange()
+                .expectStatus().isEqualTo(HttpStatus.CREATED)
+                .expectBody(Boolean.class).isEqualTo(true);
+
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
+    }
+
+    @Test
+    void addToFavorites_WhenAlreadyInFavorites_Returns200() {
+        String userId = UUID.randomUUID().toString();
+        String challengeId = UUID.randomUUID().toString();
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
+                .thenReturn(Mono.just(false));
+
+        webTestClient.post()
+                .uri("/itachallenge/api/v1/user/users/" + userId + "/favorites/" + challengeId)
+                .exchange()
+                .expectStatus().isOk()
+                .expectBody(Boolean.class).isEqualTo(false);
+
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
+    }
+
+    @Test
+    void addToFavorites_WhenUserNotExists_Returns404() {
+        String userId = UUID.randomUUID().toString();
+        String challengeId = UUID.randomUUID().toString();
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
+                .thenReturn(Mono.error(new NotFoundException("User not found")));
+
+        webTestClient.post()
+                .uri("/itachallenge/api/v1/user/users/" + userId + "/favorites/" + challengeId)
+                .exchange()
+                .expectStatus().isNotFound()
+                .expectBody(String.class).isEqualTo("User not found");
+
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
+    }
+
+    @Test
+    void addToFavorites_WhenBadFormattedId_Returns400() {
+        String userId = "invalidUuid";
+        String challengeId = "invalidUUid";
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
+                .thenReturn(Mono.error(new BadUUIDException("Error message")));
+
+        webTestClient.post()
+                .uri("/itachallenge/api/v1/user/users/" + userId + "/favorites/" + challengeId)
+                .exchange()
+                .expectStatus().isBadRequest()
+                .expectBody(String.class).isEqualTo("The provided IDs are not valid.");
+
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
+    }
+
+    @Test
+    void addToFavorites_WhenUnexpectedError_Returns500() {
+        String userId = UUID.randomUUID().toString();
+        String challengeId = UUID.randomUUID().toString();
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
+                .thenReturn(Mono.error(new Exception()));
+
+        webTestClient.post()
+                .uri("/itachallenge/api/v1/user/users/" + userId + "/favorites/" + challengeId)
+                .exchange()
+                .expectStatus().isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR)
+                .expectBody(String.class).isEqualTo("Unexpected error happened.");
+
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
+    }
+
+    @Test
     @DisplayName("GET /userinteraction/favorites/{userId} returns favorite challenges")
     void getUserFavorites_returnsFavorites() {
         UUID userId = UUID.randomUUID();
